requirements.txt
bcrypt==4.1.2
black
click
fastapi==0.110.0
filetype==1.0.7
flake8==7.0.0
isort==5.13.2
jose==1.0.0
oracledb
pandas==2.2.1
passlib==1.7.4
pydantic==2.6.3
pydantic_settings==2.2.1
pydantic_settings_yaml==0.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart
Requests
#regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
SQLAlchemy==2.0.23
toml==0.10.2
uvicorn==0.32.1
gunicorn==21.2.0
mysql-connector-python==9.2.0
mysqlclient